# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Comptes",
        "description": "Documentation de l'API du système de gestion des comptes bancaires",
        "contact": {
            "email": "contact@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://laravel-projet-gestion-comptes.onrender.com/api/v1",
            "description": "Serveur de production"
        },
        {
            "url": "/"
        }
    ],
    "paths": {
        "/comptes/{numero}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Récupère les détails d'un compte spécifique",
                "description": "Récupère les détails d'un compte spécifique",
                "operationId": "f1ffa53ac6667317a7418bbe52059a19",
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Succès"
                    },
                    "404": {
                        "description": "Compte non trouvé"
                    }
                }
            }
        },
        "/comptes": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Liste tous les comptes avec pagination et filtres",
                "description": "Lister tous les comptes",
                "operationId": "62186908ce548b721ae5b90f621e56d4",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Recherche par titulaire ou numéro de compte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filtre par type de compte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "statut",
                        "in": "query",
                        "description": "Filtre par statut de compte",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Direction du tri (asc/desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Champ pour le tri",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "solde",
                                "titulaire",
                                "type",
                                "statut",
                                "numero_compte"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Nombre d'éléments par page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des comptes récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Créer un nouveau compte bancaire",
                "description": "Créer un nouveau compte",
                "operationId": "458db376a3bd23c52102bbd05cda318b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "soldeInitial",
                                    "devise",
                                    "client"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "CHEQUE",
                                            "EPARGNE"
                                        ],
                                        "example": "cheque"
                                    },
                                    "soldeInitial": {
                                        "type": "number",
                                        "minimum": 10000,
                                        "example": 500000
                                    },
                                    "devise": {
                                        "type": "string",
                                        "enum": [
                                            "FCFA",
                                            "XOF",
                                            "USD",
                                            "EUR"
                                        ],
                                        "example": "FCFA"
                                    },
                                    "client": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "format": "uuid",
                                                "example": null,
                                                "nullable": true
                                            },
                                            "titulaire": {
                                                "type": "string",
                                                "example": "Hawa BB Wane"
                                            },
                                            "nci": {
                                                "type": "string",
                                                "example": "1234567890123",
                                                "nullable": true
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "cheikh.sy@example.com"
                                            },
                                            "telephone": {
                                                "type": "string",
                                                "example": "+221771234567"
                                            },
                                            "adresse": {
                                                "type": "string",
                                                "example": "Dakar, Sénégal"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte créé avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "660f9511-f30c-52e5-b827-557766551111"
                                                },
                                                "numeroCompte": {
                                                    "type": "string",
                                                    "example": "C00123460"
                                                },
                                                "titulaire": {
                                                    "type": "string",
                                                    "example": "Cheikh Sy"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "cheque"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "example": 500000
                                                },
                                                "devise": {
                                                    "type": "string",
                                                    "example": "FCFA"
                                                },
                                                "dateCreation": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-10-19T10:30:00Z"
                                                },
                                                "statut": {
                                                    "type": "string",
                                                    "example": "actif"
                                                },
                                                "metadata": {
                                                    "properties": {
                                                        "derniereModification": {
                                                            "type": "string",
                                                            "format": "date-time",
                                                            "example": "2025-10-19T10:30:00Z"
                                                        },
                                                        "version": {
                                                            "type": "integer",
                                                            "example": 1
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "VALIDATION_ERROR"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "Les données fournies sont invalides"
                                                },
                                                "details": {
                                                    "properties": {
                                                        "client.telephone": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "Le numéro de téléphone doit être un numéro sénégalais valide"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Comptes",
            "description": "Comptes"
        }
    ]
}